Trying to clean up an antipattern
There are a bunch of cases where I kept a list of A* and connected
A::destroyed to []{ list->remove(A); }, which is a weird antipattern.

The right way to do this is not to keep the list explicitly, but use
Qt's signals and slots so that we can inform all of the A's (that
were in the list) when something happens.

It's not all cleaned up, but this fixes many of the most egregious
examples of this pattern.